Build core to verify changes
